requirements.txt
python-telegram-bot==22.7
pytz==2024.1